make program more modular: random.hpp, and logger.hpp are moved out, CMakeLists.txt moved into main directory.
